feat: backend routes protected

+++ server/src/controllers/entry.controller.ts
@@ -2,18 +2,32 @@ import prisma from '../models/db'
 import Koa from 'koa'
 
 import { Place, newEntry, SmallEntry } from '../../server-types/types'
+import { verifyUser } from './user.controller';
 
 const postEntry = async (ctx : Koa.Context) => {
-    try {
-        const newEntry = await prisma.entry.create({
-            data : <newEntry> ctx.request.body,
-        });
+    console.log(ctx.request.body)
+    if (verifyUser(ctx.request.body.token)) {
+        try {
+            let data: newEntry = {
+                placeId: ctx.request.body.placeId,
+                authorId: ctx.request.body.authorId,
+                title: ctx.request.body.title,
+                content: ctx.request.body.content
+            }
 
-        ctx.body = newEntry;
-    } catch (err) {
-        console.log(err)
-        ctx.status = 400
-        ctx.body = 'Error: could not create new entry'
+            const newEntry = await prisma.entry.create({
+                data
+            });
+
+            ctx.body = newEntry;
+        } catch (err) {
+            console.log(err)
+            ctx.status = 400
+            ctx.body = 'Error: could not create new entry'
+        }
+    } else {
+        ctx.response.status = 401;
+        ctx.response.body = "Access denied."
     }
 }
 
@@ -103,17 +117,22 @@ const getCityEntries= async (ctx : Koa.Context) => {
 }
 
 const deleteEntry = async (ctx: Koa.Context) => {
-    try {
-        const deleteEntry = await prisma.entry.delete({
-            where: {
-            id: ctx.params.entryID,
-            },
-        })
-        ctx.body = deleteEntry;
-    } catch (err) {
-        console.log(err)
-        ctx.status = 400
-        ctx.body = 'Error: could not delete entry'
+    if (verifyUser(ctx.request.body.token)) {
+        try {
+            const deleteEntry = await prisma.entry.delete({
+                where: {
+                id: ctx.params.entryID,
+                },
+            })
+            ctx.body = deleteEntry;
+        } catch (err) {
+            console.log(err)
+            ctx.status = 400
+            ctx.body = 'Error: could not delete entry'
+        }
+    } else {
+        ctx.response.status = 401;
+        ctx.response.body = 'Access denied.'
     }
 }
